add tests for analyzerManager, combine common
